Move variable color example to separate file
